+++ changelog.md
@@ -13,6 +13,7 @@
 * Fixed unable to disable "Collide with walls damage"
 * Fixed disabling the Enchanting feature not disabling the mod's enchantments
   * Also added a config option to disable ITR's Fire Aspect and Knockback
+* Fixed disabling peaceful hunger preventing players from consuming hunger
 
 ## 3.21.1
 * Changed animal caps in different seasons