+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Requires InsaneLib 1.15.1
+
 ## 3.21.2
 * Added a way to raise the hotbar position
   * Since the Raised mod is not compatible I've done it myself